Test over Space.construct_url()
